move time-report components into feature module
