+++ src/app/app-routing.module.ts
@@ -3,10 +3,7 @@ import {Routes, RouterModule} from '@angular/router';
 import {EmployeesComponent} from './employees/employees.component';
 import {AddEmployeeComponent} from './employees/add-employee/add-employee.component';
 import {ManageEmployeeComponent} from './employees/manage-employee/manage-employee.component';
-import {TimeReportsComponent} from './time-reports/time-reports.component';
-import {MonthSummaryComponent} from './time-reports/month-summary/month-summary.component';
-import {SubmitTimeReportComponent} from './submit-time-report/submit-time-report.component';
-import {ManageTimeReportComponent} from './manage-time-report/manage-time-report.component';
+import {SubmitTimeReportComponent} from './time-reports/submit-time-report/submit-time-report.component';
 import {LoginComponent} from './core/auth/login.component';
 import {LoggedOutComponent} from './core/auth/logged-out.component';
 import {AuthGuard} from './core/auth/auth.guard';
@@ -30,22 +27,15 @@ const routes: Routes = [
   },
   {
     path: 'time-reports',
-    component: TimeReportsComponent,
+    loadChildren: () =>
+      import('./time-reports/time-reports.module').then(
+        (m) => m.TimeReportsModule
+      ),
     canActivate: [AuthGuard]
   },
   {
     path: 'employees/:employeeNumber/time-sheets',
     redirectTo: 'employees/:employeeNumber',
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'time-reports/:year/:month/summary',
-    component: MonthSummaryComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'time-reports/:id',
-    component: ManageTimeReportComponent,
     canActivate: [AuthGuard]
   },
