+++ src/app/time-reports/manage-time-report/manage-time-report.component.scss
