+++ src/app/time-reports/time-reports/month-summary/month-summary.component.scss
